+++ CHANGES.md
@@ -1,6 +1,7 @@
 ### Skin Composer Version 38 ###
 
 * Added shortcut "Ctrl+e" for export in Scene Composer.
+* Updated instructions for Custom Serializer in FreeType Font dialog to include link to FreeTypeSkin.
 * JSON exports are guaranteed to be UTF-8 formatted to prevent issues with loading unicode characters.
 * Fixed crash if setting increment on Sliders/ProgressBars to a value less than 1